requirements.txt
fastapi
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
uvicorn
pytest
pytest-asyncio
httpx
prometheus-client
slowapi
